Added new features to GUI

+++ ConsoleApp/ShowActorsWind.cs
@@ -14,6 +14,9 @@ namespace ConsoleApp
         public Label pagesLabelAll;
         private Button nextPageButton;
         private Button prevPageButton;
+        private TextField searchField;
+        private string filtervalue ="";
+        private User user;
         public ShowActorsWind()
         {
             this.Title = "List of actors"; X = 10; Y = 4; Width = Dim.Fill()-10; Height = Dim.Fill()-4;
@@ -40,23 +43,47 @@ namespace ConsoleApp
             nextPageButton = new Button("Next page"){X = Pos.Right(prevPageButton)+3, Y=Pos.Top(prevPageButton)};
             nextPageButton.Clicked += OnNextPageButClicked;
             this.Add(page,pagesLabelCur,of,pagesLabelAll,prevPageButton, nextPageButton);
+
+            searchField = new TextField()
+            {
+                X = Pos.Right(nextPageButton)+2, Y = Pos.Top(nextPageButton), Width = Dim.Fill()-5
+            };
+            searchField.TextChanging += OnSearchEnter;
+            this.Add(searchField);
+        }
+        private void OnSearchEnter(TextChangingEventArgs obj)
+        {
+            this.filtervalue = obj.NewText.ToString();
+            page = 1;
+            ShowCurrPage();
         }
         public void SetService(Service service)
         {
             this.service = service;
             ShowCurrPage();
         }
+        public void SetUser(User currUser)
+        {
+            this.user = currUser; 
+        }
         private void ShowCurrPage()
         {
             this.pagesLabelCur.Text = page.ToString();
-            int total = service.actorRepository.GetTotalPages();
+            int total = service.actorRepository.GetSearchPagesCount(filtervalue);
             this.pagesLabelAll.Text = total.ToString();
-            this.allActorsListView.SetSource(service.actorRepository.GetPage(page));
+            this.allActorsListView.SetSource(service.actorRepository.GetSearchPage(filtervalue, page));
+            this.nextPageButton.Visible = true;
+            this.prevPageButton.Visible = true;
             if(total==0)
             {
                 this.page = 0;
-                this.pagesLabelCur.Text = page.ToString();
+                this.pagesLabelAll.Text ="x";
+                this.pagesLabelCur.Text ="x";
+                this.nextPageButton.Visible = false;
+                this.prevPageButton.Visible = false;
+                this.allActorsListView.SetSource(new List<string>(){"No results"});
             }
+            Application.Refresh();
         }
         private void OnQuit()
         {
@@ -65,7 +92,7 @@ namespace ConsoleApp
         }
         private void OnNextPageButClicked()
         {
-            int totalPages = service.actorRepository.GetTotalPages();
+            int totalPages = service.actorRepository.GetSearchPagesCount(filtervalue);
             if(page>=totalPages)
                 return;
             this.page += 1;
@@ -80,41 +107,31 @@ namespace ConsoleApp
         }
         private void OnOpenActor(ListViewItemEventArgs args)
         {
-            Actor actor = (Actor)args.Value;
+            Actor actor = new Actor();
+            try{
+            actor = (Actor)args.Value;}catch{return;}
             OpenActorDialog dialog = new OpenActorDialog();
             actor.films = service.roleRepository.GetAllFilms(actor.id);
             dialog.SetActor(actor);
-            dialog.SetRepository(service.filmRepository);
+            dialog.SetUser(user);
+            dialog.SetService(service);
 
             Application.Run(dialog);
-
+            if(dialog.canceled)
+            {
+                return;
+            }
             if(dialog.deleted)
             {
                 service.actorRepository.DeleteById(actor.id);
                 service.roleRepository.DeleteActorById(actor.id);
                 MessageBox.Query("Delete actor","Actor was deleted succesfully","OK");
-                int pages = service.actorRepository.GetTotalPages();
+                int pages = service.actorRepository.GetSearchPagesCount(filtervalue);
                 if(page>pages && page >1) page += -1;
                 ShowCurrPage();
             }
             if(dialog.edited)
             {
-                Actor newAc = dialog.GetActor();
-                newAc.id = actor.id;
-                service.actorRepository.Update(newAc);
-                service.roleRepository.DeleteActorById(actor.id);
-                int[] filmsId = dialog.GetFilmsId();
-
-                if(filmsId != null)
-                {
-                    foreach(int id in filmsId)
-                    {   
-                        if(id ==0 ) continue;
-                        Role role = new Role(){actor_id = newAc.id, film_id = id};
-                        service.roleRepository.Insert(role);
-                    }
-                }
-                MessageBox.Query("Edit actor","Actor was edited succesfully","OK");
                 ShowCurrPage();
             }
         }
